Suppression des images des catégories lors de la suppression de la cat.
Blank.gif n'est pas supprimé et si l'image est utilisées dans une autre
catégorie n'on plus

+++ admin/category.php
@@ -121,12 +121,20 @@ switch ($op) {
                 redirect_header('category.php', 3, implode(',', $GLOBALS['xoopsSecurity']->getErrors()));
             }
             if ($category_Handler->delete($obj)) {
-                // A modifier car si on utilise plusieurs fois la même image, elle est détruite
-                /*$urlfile = XOOPS_UPLOAD_PATH . '/xmcontact/images/cats/' . $obj->getVar('category_logo');
-                if (is_file($urlfile)) {
-                    chmod($urlfile, 0777);
-                    unlink($urlfile);
-                }*/
+                //Del logo
+				if($obj->getVar('category_logo') != 'blank.gif'){
+					// Criteria
+					$criteria = new CriteriaCompo();
+					$criteria->add(new Criteria('category_logo', $obj->getVar('category_logo')));
+					$category_count = $category_Handler->getCount($criteria);
+					if($category_count == 0){
+						$urlfile = XOOPS_UPLOAD_PATH . '/xmcontact/images/cats/' . $obj->getVar('category_logo');
+						if (is_file($urlfile)) {
+							chmod($urlfile, 0777);
+							unlink($urlfile);
+						}
+					}
+				}
                 redirect_header('category.php', 2, _AM_XMCONTACT_REDIRECT_SAVE);
             } else {
                 xoops_error($obj->getHtmlErrors());
